add serde type wrapper SerdeType, move VotingStatus serde impl into service
