+++ src/cat-data-service/src/main.rs
@@ -1,11 +1,11 @@
 use clap::Parser;
 
 mod cli;
-// mod db;
 mod logger;
 mod service;
 mod settings;
 mod state;
+mod types;
 
 #[tokio::main]
 async fn main() -> Result<(), cli::Error> {
